Fixing runtimes in /mob/living creation/qdel.

+++ code/modules/mob/living/deity/deity_items.dm
@@ -34,11 +34,13 @@
 
 
 /mob/living/deity/Destroy()
-	for(var/cat in items_by_category)
-		var/list/L = items_by_category[cat]
-		L.Cut()
-	items_by_category.Cut()
-	for(var/i in items)
-		qdel(items[i])
-	items.Cut()
+	if(islist(items_by_category))
+		for(var/cat in items_by_category)
+			var/list/L = items_by_category[cat]
+			L.Cut()
+		items_by_category.Cut()
+	if(islist(items))
+		for(var/i in items)
+			qdel(items[i])
+		items.Cut()
 	. = ..()
